requirements.txt
python-dotenv
PyYAML
requests
websocket-client
streamlit
telebot
pandas
matplotlib